+++ src/main/java/net/darkseraphim/linecount/strategy/DefaultLineCountStrategy.java
@@ -1,4 +1,6 @@
-package net.darkseraphim.linecount;
+package net.darkseraphim.linecount.strategy;
+
+import net.darkseraphim.linecount.FileIterable;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -8,7 +10,7 @@ import java.math.BigInteger;
 public class DefaultLineCountStrategy implements LineCountStrategy {
 
     public BigInteger countLines(Path path) throws IOException {
-        FileIterable iterable = new FileIterable(path); 
+        FileIterable iterable = new FileIterable(path);
         long lines = StreamSupport.stream(iterable.spliterator(), false) 
                                   .filter(c -> c == '\n') 
                                   .count();
